Updating Manifesto
Focusing in the MISSION instead of a solution proposition as it can be "n" different types collaborating to the same purpose. It shall not guide through a solution, but instead, a purpose having multiple connected solutions in the same contract. 
It defines the interfaces, and how to interact with it.